API_Vercel_Setup/requirements.txt
fastapi
google-genai
python-multipart
pillow
